Export fixed, Redirection da fixare

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: agenoves <agenoves@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/10 14:21:55 by agenoves          #+#    #+#             */
-/*   Updated: 2022/01/10 16:00:04 by agenoves         ###   ########.fr       */
+/*   Updated: 2022/10/14 12:25:56 by agenoves         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
